Codegen Ouath, Fix path
