--- hubspot/auth/oauth/__init__.py
@@ -1,39 +1,0 @@
-# coding: utf-8
-
-# flake8: noqa
-
-"""
-    Auth Oauth
-
-    No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)  # noqa: E501
-
-    The version of the OpenAPI document: v1
-    Generated by: https://openapi-generator.tech
-"""
-
-
-from __future__ import absolute_import
-
-__version__ = "1.0.0"
-
-# import apis into sdk package
-from hubspot.auth.oauth.api.access_tokens_api import AccessTokensApi
-from hubspot.auth.oauth.api.refresh_tokens_api import RefreshTokensApi
-from hubspot.auth.oauth.api.tokens_api import TokensApi
-
-# import ApiClient
-from hubspot.auth.oauth.api_client import ApiClient
-from hubspot.auth.oauth.configuration import Configuration
-from hubspot.auth.oauth.exceptions import OpenApiException
-from hubspot.auth.oauth.exceptions import ApiTypeError
-from hubspot.auth.oauth.exceptions import ApiValueError
-from hubspot.auth.oauth.exceptions import ApiKeyError
-from hubspot.auth.oauth.exceptions import ApiAttributeError
-from hubspot.auth.oauth.exceptions import ApiException
-
-# import models into sdk package
-from hubspot.auth.oauth.models.access_token_info_response import AccessTokenInfoResponse
-from hubspot.auth.oauth.models.error import Error
-from hubspot.auth.oauth.models.error_detail import ErrorDetail
-from hubspot.auth.oauth.models.refresh_token_info_response import RefreshTokenInfoResponse
-from hubspot.auth.oauth.models.token_response_if import TokenResponseIF
