--- hubspot/auth/oauth/api/__init__.py
@@ -1,8 +1,0 @@
-from __future__ import absolute_import
-
-# flake8: noqa
-
-# import apis into api package
-from hubspot.auth.oauth.api.access_tokens_api import AccessTokensApi
-from hubspot.auth.oauth.api.refresh_tokens_api import RefreshTokensApi
-from hubspot.auth.oauth.api.tokens_api import TokensApi
