--- hubspot/auth/oauth/models/__init__.py
@@ -1,21 +1,0 @@
-# coding: utf-8
-
-# flake8: noqa
-"""
-    Auth Oauth
-
-    No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)  # noqa: E501
-
-    The version of the OpenAPI document: v1
-    Generated by: https://openapi-generator.tech
-"""
-
-
-from __future__ import absolute_import
-
-# import models into model package
-from hubspot.auth.oauth.models.access_token_info_response import AccessTokenInfoResponse
-from hubspot.auth.oauth.models.error import Error
-from hubspot.auth.oauth.models.error_detail import ErrorDetail
-from hubspot.auth.oauth.models.refresh_token_info_response import RefreshTokenInfoResponse
-from hubspot.auth.oauth.models.token_response_if import TokenResponseIF
