+++ hubspot/oauth/__init__.py
@@ -1,0 +1,39 @@
+# coding: utf-8
+
+# flake8: noqa
+
+"""
+    Auth Oauth
+
+    No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)  # noqa: E501
+
+    The version of the OpenAPI document: v1
+    Generated by: https://openapi-generator.tech
+"""
+
+
+from __future__ import absolute_import
+
+__version__ = "1.0.0"
+
+# import apis into sdk package
+from hubspot.oauth.api.access_tokens_api import AccessTokensApi
+from hubspot.oauth.api.refresh_tokens_api import RefreshTokensApi
+from hubspot.oauth.api.tokens_api import TokensApi
+
+# import ApiClient
+from hubspot.oauth.api_client import ApiClient
+from hubspot.oauth.configuration import Configuration
+from hubspot.oauth.exceptions import OpenApiException
+from hubspot.oauth.exceptions import ApiTypeError
+from hubspot.oauth.exceptions import ApiValueError
+from hubspot.oauth.exceptions import ApiKeyError
+from hubspot.oauth.exceptions import ApiAttributeError
+from hubspot.oauth.exceptions import ApiException
+
+# import models into sdk package
+from hubspot.oauth.models.access_token_info_response import AccessTokenInfoResponse
+from hubspot.oauth.models.error import Error
+from hubspot.oauth.models.error_detail import ErrorDetail
+from hubspot.oauth.models.refresh_token_info_response import RefreshTokenInfoResponse
+from hubspot.oauth.models.token_response_if import TokenResponseIF
