+++ hubspot/oauth/api/tokens_api.py
@@ -17,11 +17,11 @@ import re  # noqa: F401
 # python 2 and python 3 compatibility library
 import six
 
-from hubspot.auth.oauth.api_client import ApiClient
-from hubspot.auth.oauth.exceptions import ApiTypeError, ApiValueError  # noqa: F401
+from hubspot.oauth.api_client import ApiClient
+from hubspot.oauth.exceptions import ApiTypeError, ApiValueError  # noqa: F401
 
 
-class AccessTokensApi(object):
+class TokensApi(object):
     """NOTE: This class is auto generated by OpenAPI Generator
     Ref: https://openapi-generator.tech
 
@@ -33,17 +33,27 @@ class AccessTokensApi(object):
             api_client = ApiClient()
         self.api_client = api_client
 
-    def get(self, token, **kwargs):  # noqa: E501
-        """get  # noqa: E501
+    def create(self, **kwargs):  # noqa: E501
+        """create  # noqa: E501
 
         This method makes a synchronous HTTP request by default. To make an
         asynchronous HTTP request, please pass async_req=True
 
-        >>> thread = api.get(token, async_req=True)
+        >>> thread = api.create(async_req=True)
         >>> result = thread.get()
 
-        :param token: (required)
-        :type token: str
+        :param grant_type:
+        :type grant_type: str
+        :param code:
+        :type code: str
+        :param redirect_uri:
+        :type redirect_uri: str
+        :param client_id:
+        :type client_id: str
+        :param client_secret:
+        :type client_secret: str
+        :param refresh_token:
+        :type refresh_token: str
         :param async_req: Whether to execute the request asynchronously.
         :type async_req: bool, optional
         :param _preload_content: if False, the urllib3.HTTPResponse object will
@@ -57,22 +67,32 @@ class AccessTokensApi(object):
         :return: Returns the result object.
                  If the method is called asynchronously,
                  returns the request thread.
-        :rtype: AccessTokenInfoResponse
+        :rtype: TokenResponseIF
         """
         kwargs["_return_http_data_only"] = True
-        return self.get_with_http_info(token, **kwargs)  # noqa: E501
+        return self.create_with_http_info(**kwargs)  # noqa: E501
 
-    def get_with_http_info(self, token, **kwargs):  # noqa: E501
-        """get  # noqa: E501
+    def create_with_http_info(self, **kwargs):  # noqa: E501
+        """create  # noqa: E501
 
         This method makes a synchronous HTTP request by default. To make an
         asynchronous HTTP request, please pass async_req=True
 
-        >>> thread = api.get_with_http_info(token, async_req=True)
+        >>> thread = api.create_with_http_info(async_req=True)
         >>> result = thread.get()
 
-        :param token: (required)
-        :type token: str
+        :param grant_type:
+        :type grant_type: str
+        :param code:
+        :type code: str
+        :param redirect_uri:
+        :type redirect_uri: str
+        :param client_id:
+        :type client_id: str
+        :param client_secret:
+        :type client_secret: str
+        :param refresh_token:
+        :type refresh_token: str
         :param async_req: Whether to execute the request asynchronously.
         :type async_req: bool, optional
         :param _return_http_data_only: response data without head status code
@@ -94,28 +114,23 @@ class AccessTokensApi(object):
         :return: Returns the result object.
                  If the method is called asynchronously,
                  returns the request thread.
-        :rtype: tuple(AccessTokenInfoResponse, status_code(int), headers(HTTPHeaderDict))
+        :rtype: tuple(TokenResponseIF, status_code(int), headers(HTTPHeaderDict))
         """
 
         local_var_params = locals()
 
-        all_params = ["token"]
+        all_params = ["grant_type", "code", "redirect_uri", "client_id", "client_secret", "refresh_token"]
         all_params.extend(["async_req", "_return_http_data_only", "_preload_content", "_request_timeout", "_request_auth", "_content_type", "_headers"])
 
         for key, val in six.iteritems(local_var_params["kwargs"]):
             if key not in all_params:
-                raise ApiTypeError("Got an unexpected keyword argument '%s'" " to method get" % key)
+                raise ApiTypeError("Got an unexpected keyword argument '%s'" " to method create" % key)
             local_var_params[key] = val
         del local_var_params["kwargs"]
-        # verify the required parameter 'token' is set
-        if self.api_client.client_side_validation and local_var_params.get("token") is None:  # noqa: E501
-            raise ApiValueError("Missing the required parameter `token` when calling `get`")  # noqa: E501
 
         collection_formats = {}
 
         path_params = {}
-        if "token" in local_var_params:
-            path_params["token"] = local_var_params["token"]  # noqa: E501
 
         query_params = []
 
@@ -123,21 +138,38 @@ class AccessTokensApi(object):
 
         form_params = []
         local_var_files = {}
+        if "grant_type" in local_var_params:
+            form_params.append(("grant_type", local_var_params["grant_type"]))  # noqa: E501
+        if "code" in local_var_params:
+            form_params.append(("code", local_var_params["code"]))  # noqa: E501
+        if "redirect_uri" in local_var_params:
+            form_params.append(("redirect_uri", local_var_params["redirect_uri"]))  # noqa: E501
+        if "client_id" in local_var_params:
+            form_params.append(("client_id", local_var_params["client_id"]))  # noqa: E501
+        if "client_secret" in local_var_params:
+            form_params.append(("client_secret", local_var_params["client_secret"]))  # noqa: E501
+        if "refresh_token" in local_var_params:
+            form_params.append(("refresh_token", local_var_params["refresh_token"]))  # noqa: E501
 
         body_params = None
         # HTTP header `Accept`
         header_params["Accept"] = self.api_client.select_header_accept(["application/json", "*/*"])  # noqa: E501
 
+        # HTTP header `Content-Type`
+        content_types_list = local_var_params.get("_content_type", self.api_client.select_header_content_type(["application/x-www-form-urlencoded"], "POST", body_params))  # noqa: E501
+        if content_types_list:
+            header_params["Content-Type"] = content_types_list
+
         # Authentication setting
         auth_settings = []  # noqa: E501
 
         response_types_map = {
-            200: "AccessTokenInfoResponse",
+            200: "TokenResponseIF",
         }
 
         return self.api_client.call_api(
-            "/oauth/v1/access-tokens/{token}",
-            "GET",
+            "/oauth/v1/token",
+            "POST",
             path_params,
             query_params,
             header_params,
